+++ pkg/codegen/internal/test/testdata/simple-resource-schema/go/example/resource.go
@@ -76,15 +76,15 @@ type ResourceInput interface {
 	ToResourceOutputWithContext(ctx context.Context) ResourceOutput
 }
 
-func (Resource) ElementType() reflect.Type {
-	return reflect.TypeOf((*Resource)(nil)).Elem()
+func (*Resource) ElementType() reflect.Type {
+	return reflect.TypeOf((*Resource)(nil))
 }
 
-func (i Resource) ToResourceOutput() ResourceOutput {
+func (i *Resource) ToResourceOutput() ResourceOutput {
 	return i.ToResourceOutputWithContext(context.Background())
 }
 
-func (i Resource) ToResourceOutputWithContext(ctx context.Context) ResourceOutput {
+func (i *Resource) ToResourceOutputWithContext(ctx context.Context) ResourceOutput {
 	return pulumi.ToOutputWithContext(ctx, i).(ResourceOutput)
 }
 
@@ -93,7 +93,7 @@ type ResourceOutput struct {
 }
 
 func (ResourceOutput) ElementType() reflect.Type {
-	return reflect.TypeOf((*ResourceOutput)(nil)).Elem()
+	return reflect.TypeOf((*Resource)(nil))
 }
 
 func (o ResourceOutput) ToResourceOutput() ResourceOutput {